requirements.txt
hupper==1.0
PasteDeploy==1.5.2
plaster==0.5
plaster-pastedeploy==0.4.1
pyramid==1.9.1
repoze.lru==0.6
SQLAlchemy==1.1.13
transaction==2.1.2
translationstring==1.3
venusian==1.1.0
WebOb==1.7.3
zope.deprecation==4.3.0
zope.interface==4.4.2
zope.sqlalchemy==0.7.7
